requirements.txt
pip>=25.1
setuptools>=75.0
wheel

certifi==2023.7.22
charset-normalizer==2.0.11
Django==4.1.13
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt>=5.3.1
idna==3.7
gunicorn==22.0.0
Pillow==10.4.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sslcommerz-lib==1.0
urllib3==1.26.18
virtualenv==20.14.1
whitenoise==6.2.0
widgetsnbextension==3.5.2
xhtml2pdf==0.2.8

google-generativeai